+++ src/components/income.jsx
@@ -3,15 +3,7 @@ import { withStyles } from "@material-ui/core/styles";
 import AddIncome from "./addIncome";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
-import MaterialTable from "material-table";
-import { connect } from "react-redux";
-import {
-	getIncome,
-	addIncome,
-	deleteIncome,
-	updateIncome
-} from "../actions/income";
-import moment from "moment";
+import MaterialIncomeTable from "./materialIncomeTable";
 
 const styles = theme => ({
 	root: {
@@ -35,97 +27,25 @@ const styles = theme => ({
 });
 
 class Income extends Component {
-	state = {
-		columns: [
-			{ title: "Date", field: "date", type: "date" },
-			{ title: "Payer", field: "payer" },
-			{ title: "Income", field: "income", type: "numeric" },
-			{ title: "Notes", field: "notes" }
-		],
-		data: this.props.income
-	};
-
-	componentDidMount() {
-		getIncome();
-	}
-
 	render() {
 		const { classes } = this.props;
-		console.log(this.props);
-		return (
-			<div>
-				<Grid container className={classes.root}>
-					<Grid item xs={4} className={classes.income}>
-						<Paper className={classes.paper}>
-							<AddIncome />
-						</Paper>
-					</Grid>
 
-					<Grid item xs={7} className={classes.table}>
-						<Paper className={classes.paper}>
-							<MaterialTable
-								title="Manage Income"
-								columns={this.state.columns}
-								data={this.state.data}
-								editable={{
-									onRowAdd: newData =>
-										new Promise(resolve => {
-											setTimeout(() => {
-												resolve();
-												const data = [...this.state.data];
-												newData.date = moment(newData.date).format(
-													"MM/DD/YYYY"
-												);
-												data.push(newData);
-												this.setState({ data });
-												this.props.addIncome(newData);
-											}, 600);
-										}),
-									onRowUpdate: (newData, oldData) =>
-										new Promise(resolve => {
-											setTimeout(() => {
-												resolve();
-												const data = [...this.state.data];
-												data[data.indexOf(oldData)] = newData;
-												this.setState({ data });
-												this.props.updateIncome(newData);
-											}, 600);
-										}),
-									onRowDelete: oldData =>
-										new Promise(resolve => {
-											setTimeout(() => {
-												resolve();
-												const data = [...this.state.data];
-												data.splice(data.indexOf(oldData), 1);
-												this.setState({ data });
-												console.log(oldData.id);
-												this.props.deleteIncome(oldData.id);
-											}, 600);
-										})
-								}}
-							/>
-						</Paper>
-					</Grid>
+		return (
+			<Grid container className={classes.root}>
+				<Grid item xs={4} className={classes.income}>
+					<Paper className={classes.paper}>
+						<AddIncome />
+					</Paper>
 				</Grid>
-			</div>
+
+				<Grid item xs={7} className={classes.table}>
+					<Paper className={classes.paper}>
+						<MaterialIncomeTable />
+					</Paper>
+				</Grid>
+			</Grid>
 		);
 	}
 }
-const mapStateToProps = state => {
-	return {
-		income: state.income.income
-	};
-};
-const mapDispatchToProps = dispatch => {
-	return {
-		getIncome: dispatch(getIncome()),
-		addIncome: income => dispatch(addIncome(income)),
-		deleteIncome: incomeID => dispatch(deleteIncome(incomeID)),
-		updateIncome: income => dispatch(updateIncome(income))
-	};
-};
 
-export default connect(
-	mapStateToProps,
-	mapDispatchToProps
-)(withStyles(styles)(Income));
+export default withStyles(styles)(Income);
