+++ src/components/materialIncomeTable.jsx
@@ -1,0 +1,88 @@
+import React, { Component } from "react";
+import MaterialTable from "material-table";
+import { connect } from "react-redux";
+import {
+	getIncome,
+	addIncome,
+	deleteIncome,
+	updateIncome
+} from "../actions/income";
+import moment from "moment";
+
+class MaterialIncomeTable extends Component {
+	state = {
+		columns: [
+			{ title: "Date", field: "date", type: "date" },
+			{ title: "Payer", field: "payer" },
+			{ title: "Income", field: "income", type: "numeric" },
+			{ title: "Notes", field: "notes" }
+		],
+		data: this.props.income
+	};
+
+	componentDidMount() {
+		getIncome();
+	}
+
+	render() {
+		return (
+			<MaterialTable
+				title="Manage Income"
+				columns={this.state.columns}
+				data={this.state.data}
+				editable={{
+					onRowAdd: newData =>
+						new Promise(resolve => {
+							setTimeout(() => {
+								resolve();
+								const data = [...this.state.data];
+								newData.date = moment(newData.date).format("MM/DD/YYYY");
+								data.push(newData);
+								this.setState({ data });
+								this.props.addIncome(newData);
+							}, 600);
+						}),
+					onRowUpdate: (newData, oldData) =>
+						new Promise(resolve => {
+							setTimeout(() => {
+								resolve();
+								const data = [...this.state.data];
+								data[data.indexOf(oldData)] = newData;
+								this.setState({ data });
+								this.props.updateIncome(newData);
+							}, 600);
+						}),
+					onRowDelete: oldData =>
+						new Promise(resolve => {
+							setTimeout(() => {
+								resolve();
+								const data = [...this.state.data];
+								data.splice(data.indexOf(oldData), 1);
+								this.setState({ data });
+								console.log(oldData.id);
+								this.props.deleteIncome(oldData.id);
+							}, 600);
+						})
+				}}
+			/>
+		);
+	}
+}
+const mapStateToProps = state => {
+	return {
+		income: state.income.income
+	};
+};
+const mapDispatchToProps = dispatch => {
+	return {
+		getIncome: dispatch(getIncome()),
+		addIncome: income => dispatch(addIncome(income)),
+		deleteIncome: incomeID => dispatch(deleteIncome(incomeID)),
+		updateIncome: income => dispatch(updateIncome(income))
+	};
+};
+
+export default connect(
+	mapStateToProps,
+	mapDispatchToProps
+)(MaterialIncomeTable);
